Refactoring observation code into a separate submodule.
